% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\title{Do simulation based calibration.
Returns an S3 object, with plot method showing some stuff . . .}
\usage{
calibrate(model, pior_listb)
}
\arguments{
\item{returns}{the prior specification.}

\item{prior_list:}{expects a list with the following:
alpha: a postive real numnber (first parameter for beta prior on tau).
beta: a positive real number (second parameter for beta prior on tau).
hyperparams:  .  . . .}
}
\value{
an S3 object . . .
}
\description{
Do simulation based calibration.
Returns an S3 object, with plot method showing some stuff . . .
}
