% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_modes.R
\name{post_mode_delta}
\alias{post_mode_delta}
\title{Compute the posterior mode of \eqn{\delta}}
\usage{
post_mode_delta(x)
}
\arguments{
\item{x}{Tibble of draws (filtered to a fixed r), with a \code{delta_test} list-column.}
}
\value{
\eqn{v \times r} matrix of the modal sparsity pattern.
}
\description{
Finds the most frequently visited sparsity pattern across a set of draws
with the same number of factors r.
}
