% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc.R
\name{print.sbc_result}
\alias{print.sbc_result}
\title{Print an sbc_result object}
\usage{
\method{print}{sbc_result}(x, ...)
}
\arguments{
\item{x}{An \code{sbc_result} object.}

\item{...}{Ignored.}
}
\description{
Print an sbc_result object
}
