% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_pivots.R
\name{switch_pivots_move}
\alias{switch_pivots_move}
\title{Switch pivots move}
\usage{
switch_pivots_move(
  delta,
  pivots,
  factors,
  y,
  theta,
  alpha,
  beta,
  hyperparams,
  j
)
}
\arguments{
\item{delta}{\eqn{v \times q} binary sparsity matrix \eqn{\delta}.}

\item{pivots}{Length-q pivot vector.}

\item{factors}{\eqn{q \times N} factor matrix \eqn{F}.}

\item{y}{\eqn{v \times N} data matrix \eqn{Y}.}

\item{theta}{Length-q column shrinkage vector \eqn{\theta}.}

\item{alpha}{Length-v shape parameters \eqn{\alpha}.}

\item{beta}{Length-v rate parameters \eqn{\beta}.}

\item{hyperparams}{List with \code{aH} and \code{bH}.}

\item{j}{Column index to update.}
}
\value{
List with updated \code{delta} and \code{pivots}.
}
\description{
Proposes swapping indicators \eqn{\delta_{i,j} \leftrightarrow \delta_{i,l}} over the rows where they
differ, for a randomly selected second column l. Accepted via MH.
}
